+++ lua/neotest-dotnet/tree-sitter/mstest-queries.lua
@@ -40,8 +40,8 @@ function M.get_queries(custom_attributes)
           name: (identifier) @attribute_name (#eq? @attribute_name "TestClass")
         )
       )
-      name: (identifier) @namespace.name
-    ) @namespace.definition
+      name: (identifier) @class.name
+    ) @class.definition
 
     ;; Matches test methods
     (method_declaration