+++ tests/nunit/specs/test_simple.cs
@@ -1,0 +1,17 @@
+using NUnit.Framework;
+
+namespace NUnitSamples;
+
+public class SingleTests
+{
+	[SetUp]
+	public void Setup()
+	{
+	}
+
+	[Test]
+	public void Test1()
+	{
+		Assert.Pass();
+	}
+}