+++ tests/xunit/discover_positions/theory_attribute_spec.lua
@@ -14,8 +14,8 @@ describe("discover_positions", function()
   })
 
   async.it("should discover tests with inline parameters", function()
-    local spec_file = "./tests/xunit/specs/basic_tests.cs"
-    local spec_file_name = "basic_tests.cs"
+    local spec_file = "./tests/xunit/specs/theory_and_fact_mixed.cs"
+    local spec_file_name = "theory_and_fact_mixed.cs"
     local positions = plugin.discover_positions(spec_file):to_list()
 
     local function get_expected_output(file_path, file_name)
@@ -30,6 +30,7 @@ describe("discover_positions", function()
         {
           {
             id = file_path .. "::xunit.testproj1",
+            is_class = false,
             name = "xunit.testproj1",
             path = file_path,
             range = { 0, 0, 17, 1 },
@@ -38,6 +39,7 @@ describe("discover_positions", function()
           {
             {
               id = file_path .. "::xunit.testproj1::UnitTest1",
+              is_class = true,
               name = "UnitTest1",
               path = file_path,
               range = { 2, 0, 17, 1 },
@@ -46,6 +48,7 @@ describe("discover_positions", function()
             {
               {
                 id = file_path .. "::xunit.testproj1::UnitTest1::Test1",
+                is_class = false,
                 name = "Test1",
                 path = file_path,
                 range = { 4, 1, 8, 2 },
@@ -55,6 +58,7 @@ describe("discover_positions", function()
             {
               {
                 id = file_path .. "::xunit.testproj1::UnitTest1::Test2",
+                is_class = false,
                 name = "Test2",
                 path = file_path,
                 range = { 10, 1, 16, 2 },
@@ -63,6 +67,7 @@ describe("discover_positions", function()
               {
                 {
                   id = file_path .. "::xunit.testproj1::UnitTest1::Test2(a: 1)",
+                  is_class = false,
                   name = "Test2(a: 1)",
                   path = file_path,
                   range = { 11, 12, 11, 15 },
@@ -72,6 +77,7 @@ describe("discover_positions", function()
               {
                 {
                   id = file_path .. "::xunit.testproj1::UnitTest1::Test2(a: 2)",
+                  is_class = false,
                   name = "Test2(a: 2)",
                   path = file_path,
                   range = { 12, 12, 12, 15 },
@@ -102,6 +108,7 @@ describe("discover_positions", function()
       {
         {
           id = "./tests/xunit/specs/block_scoped_namespace.cs::xunit.testproj1",
+          is_class = false,
           name = "xunit.testproj1",
           path = "./tests/xunit/specs/block_scoped_namespace.cs",
           range = { 0, 0, 10, 1 },
@@ -110,6 +117,7 @@ describe("discover_positions", function()
         {
           {
             id = "./tests/xunit/specs/block_scoped_namespace.cs::xunit.testproj1::UnitTest1",
+            is_class = true,
             name = "UnitTest1",
             path = "./tests/xunit/specs/block_scoped_namespace.cs",
             range = { 2, 1, 9, 2 },
@@ -118,6 +126,7 @@ describe("discover_positions", function()
           {
             {
               id = "./tests/xunit/specs/block_scoped_namespace.cs::xunit.testproj1::UnitTest1::Test1",
+              is_class = false,
               name = "Test1",
               path = "./tests/xunit/specs/block_scoped_namespace.cs",
               range = { 4, 2, 8, 3 },
@@ -129,9 +138,4 @@ describe("discover_positions", function()
     }
     assert.same(positions, expected_positions)
   end)
-
-  -- TODO:
-  -- 1. Write tests for non-inline parameterized tests
-  -- 2. Write tests for nested namespaces
-  -- 3. Write tests for nested classes
 end)